# perf(runtime): cache cwd lookups