This caches CWD lookups, `Deno.cwd()` is ~5x faster and shaves off ~20% of `read_128k_sync`.

This is a first PR of a series to optimize sync file reads, optimizing `cwd` not only helps `Deno.cwd()` but it also optimizes all file IO permission checks

## Bench results

```
# Before
cwd:                 	n = 50000, dt = 0.389s, r = 128535/s, t = 7780ns/op
read_128k_sync:      	n = 50000, dt = 1.922s, r = 26015/s, t = 38440ns/op

# After
cwd:                 	n = 50000, dt = 0.073s, r = 684932/s, t = 1460ns/op
read_128k_sync:      	n = 50000, dt = 1.591s, r = 31427/s, t = 31819ns/op
```