@AaronO 2021-09-13
> Does not seem sound as cwd can be changed from any thread in a worker.

Yep I should have explicited the assumption here, I'll change it to use a `Mutex` or `RwLock` should have roughly same performance and there shouldn't be significant contention in any real world programs