@AaronO 2021-09-13
> I'm worried about this change - it might introduce some odd bugs if somehow the cwd changes outside of op_chdir.
> 
> I don't understand why caching cwd would effect readFile perf at all. readFile should not use cwd even once let alone multiple times.

See:
> optimizing `cwd` not only helps `Deno.cwd()` but it also optimizes all file IO permission checks
since file IO permission checks need to resolve the full file path, these cwd related calls accounted for ~20% of the time spent  in the `read_128k_sync` bench, which is quite significant overhead (making us a few times slower than NodeJS for basic file IO)