@AaronO 2021-09-13
> I'm worried about this change - it might introduce some odd bugs if somehow the cwd changes outside of op_chdir.
> 
> I don't understand why caching cwd would effect readFile perf at all. readFile should not use cwd even once let alone multiple times.

In addition, I think it's a fair assumption that the CLI should never change the CWD and thus it will only be changed by user code executed in the runtime, a `lazy_static` `Mutex` or `RwLock` would the cross-thread/worker soundness casper mentioned