@bartlomieju 2021-10-30
I'm also not in favor of this change. I think we should close this PR.